implement save,load,autosave, and buttonpressing
implemented save, load, autosave commands and make it press buttons when
it's supposed to. i renamed parser.py to parser_please_work.py because
there was some weird error where it said that a function in parser
wasn't in parser so i just renamed it to parser_please_work and it
worked. no clue why that happened but i'm not questioning it